requirements.txt
pyserial
signalrcore
turfpy==0.0.7
python-dotenv